requirements.txt
asgiref==3.3.1
boto3==1.17.3
botocore==1.20.3
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
django-storages==1.11.1
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
packaging==20.8
Pillow==8.1.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
